Use proper variable types to suppress warning (#1436)

+++ ext/RMagick/rmagick.h
@@ -1164,7 +1164,7 @@ extern void   rm_check_ary_len(VALUE, long);
 extern VALUE  rm_check_ary_type(VALUE ary);
 extern Image *rm_check_destroyed(VALUE);
 extern Image *rm_check_frozen(VALUE);
-extern char  *rm_str2cstr(VALUE, long *);
+extern char  *rm_str2cstr(VALUE, size_t *);
 extern int    rm_check_num2dbl(VALUE);
 extern double rm_fuzz_to_dbl(VALUE);
 extern Quantum rm_app2quantum(VALUE);
